list_style/refactor layout
