+++ src/container/main/components/index.style.tsx
@@ -15,3 +15,10 @@ export const FooterStyled = styled(AppBar)`
     position: relative;
     bottom: 0;
 `;
+
+export const LayoutStyled = styled.div`
+    background-color: rgba(0, 0, 0, 0.54);
+    min-height: 100vh;
+    display: flex;
+    flex-direction: column;
+`;
