--- src/container/main/index.style.tsx
@@ -1,8 +1,0 @@
-import styled from '@emotion/styled';
-
-export const BodyDivStyled = styled.div`
-    background-color: rgba(0, 0, 0, 0.54);
-    min-height: 100vh;
-    display: flex;
-    flex-direction: column;
-`;
