+++ src/container/main/index.tsx
@@ -1,14 +1,9 @@
 import React from 'react';
 import { RouterProvider } from 'react-router-dom';
 import { router } from './router';
-import BodyDiv from './body-div';
 
 const Main = (): React.ReactElement => {
-    return (
-        <BodyDiv>
-            <RouterProvider router={router} />
-        </BodyDiv>
-    );
+    return <RouterProvider router={router} />;
 };
 
 export default Main;
