+++ src/container/main/router.tsx
@@ -3,24 +3,24 @@ import { createBrowserRouter } from 'react-router-dom';
 import ListPage from '../list';
 import DetailPage from '../detail';
 import FormPage from '../form';
-import { getNavigationsValue } from '@brojs/cli';
-import Layout from './components/layout';
+import { getNavigationValue } from '@brojs/cli';
+import Layout from './components';
 
 export const router = createBrowserRouter([
     {
-        path: getNavigationsValue('sber_web.main'),
+        path: getNavigationValue('sber_web.main'),
         element: <Layout />,
         children: [
             {
-                path: getNavigationsValue('sber_web.main'),
+                path: getNavigationValue('sber_web.main'),
                 element: <ListPage />
             },
             {
-                path: getNavigationsValue('sber_web.detail'),
+                path: getNavigationValue('sber_web.detail'),
                 element: <DetailPage />
             },
             {
-                path: getNavigationsValue('sber_web.edit'),
+                path: getNavigationValue('sber_web.edit'),
                 element: <FormPage />
             }
         ]
